fix: use f64 for temperature to avoid floating-point precision loss in JSON serialization

+++ crates/test-harness/src/mocks/llm.rs
@@ -131,7 +131,7 @@ impl LLMProvider for MockLLMProvider {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> bendclaw::base::Result<LLMResponse> {
         let turn = self.next_turn();
         match turn {
@@ -176,7 +176,7 @@ impl LLMProvider for MockLLMProvider {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> ResponseStream {
         let turn = self.next_turn();
         let (writer, stream) = ResponseStream::channel(16);
