+++ tests/unit/kernel/run/persister.rs
@@ -35,7 +35,7 @@ impl LLMProvider for PricingLLM {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> bendclaw::base::Result<LLMResponse> {
         Err(ErrorCode::internal("not used in persister tests"))
     }
@@ -45,7 +45,7 @@ impl LLMProvider for PricingLLM {
         _model: &str,
         _messages: &[ChatMessage],
         _tools: &[ToolSchema],
-        _temperature: f32,
+        _temperature: f64,
     ) -> ResponseStream {
         let (_writer, stream) = ResponseStream::channel(1);
         stream
@@ -59,7 +59,7 @@ impl LLMProvider for PricingLLM {
         "mock"
     }
 
-    fn default_temperature(&self) -> f32 {
+    fn default_temperature(&self) -> f64 {
         0.0
     }
 }
